requirements.txt
streamlit==1.51.0
pandas==2.2.2
numpy==1.26.4
gspread==6.2.1
google-auth==2.35.0
